[#79] feat: surface iCloud-backup reassurance + consolidate data controls
Tier 0 of #79: rather than build CloudKit, turn the iCloud device-backup
coverage the app already gets for free into a visible trust signal. The
Core Data store lives in the App Group container and is included in the
user's iCloud device backup; this makes that legible and tidies the surface.

- New DataSettingsView sub-view: device-first "Backup" reassurance section
  plus the relocated Export/Import controls.
- SettingsView: "Data" section now links to the sub-view; export/import
  state and modifiers moved out. The Contacts-import success banner and its
  state intentionally stay (they belong to the People "Add from Contacts"
  flow, not file import).
- Reconcile About-footer copy: "never leaves your phone" -> "lives on your
  device and is never sent to us" (honest once iCloud backup is acknowledged).

No CloudKit, no Core Data model change, no new persistence. Build clean
(0 warnings); 666 unit tests green.
